adding all features for required orderbook

+++ contracts/interfaces/IAbstractOrderBook.sol
@@ -3,8 +3,8 @@
 pragma solidity ^0.8.0; 
 
 interface IAbstractOrderBook {
-    function addBuyOrder(uint96 _price, uint256 size) external;
-    function addSellOrder(uint96 _price, uint256 size) external;
+    function addBuyOrder(uint96 _price, uint128 size) external;
+    function addSellOrder(uint96 _price, uint128 size) external;
     function cancelSellOrder(uint256 _orderId) external;
     function cancelBuyOrder(uint256 _orderId) external;
 }
